fix: include custom auth in each classes
