+++ eodag/plugins/search/build_search_result.py
@@ -1264,66 +1264,3 @@ class MeteoblueSearch(ECMWFSearch):
         :return: formatted query params and encode query string
         """
         return QueryStringSearch.build_query_string(self, product_type, query_dict)
-
-
-class WekeoECMWFSearch(ECMWFSearch):
-    """
-    WekeoECMWFSearch search plugin.
-
-    This plugin, which inherits from :class:`~eodag.plugins.search.build_search_result.ECMWFSearch`,
-    performs a POST request and uses its result to build a single :class:`~eodag.api.search_result.SearchResult`
-    object. In contrast to ECMWFSearch or MeteoblueSearch, the products are only build with information
-    returned by the provider.
-
-    The available configuration parameters are inherited from parent classes, with some a particularity
-    for pagination for this plugin.
-
-    :param provider: An eodag providers configuration dictionary
-    :param config: Search plugin configuration:
-
-        * :attr:`~eodag.config.PluginConfig.pagination` (:class:`~eodag.config.PluginConfig.Pagination`)
-          (**mandatory**): The configuration of how the pagination is done on the provider. For
-          this plugin it has the node:
-
-          * :attr:`~eodag.config.PluginConfig.Pagination.next_page_query_obj` (``str``): The
-            additional parameters needed to perform search. These parameters won't be included in
-            the result. This must be a json dict formatted like ``{{"foo":"bar"}}`` because it
-            will be passed to a :meth:`str.format` method before being loaded as json.
-    """
-
-    def normalize_results(
-        self, results: RawSearchResult, **kwargs: Any
-    ) -> list[EOProduct]:
-        """Build :class:`~eodag.api.product._product.EOProduct` from provider result
-
-        :param results: Raw provider result as single dict in list
-        :param kwargs: Search arguments
-        :returns: list of single :class:`~eodag.api.product._product.EOProduct`
-        """
-
-        # formating of orderLink requires access to the productType value.
-        results.data = [
-            {**result, **results.product_type_def_params} for result in results
-        ]
-
-        normalized = QueryStringSearch.normalize_results(self, results, **kwargs)
-
-        if not normalized:
-            return normalized
-
-        query_params_encoded = quote_plus(orjson.dumps(results.query_params))
-        for product in normalized:
-            properties = {**product.properties, **results.query_params}
-            properties["_dc_qs"] = query_params_encoded
-            product.properties = {ecmwf_format(k): v for k, v in properties.items()}
-
-        return normalized
-
-    def do_search(self, *args: Any, **kwargs: Any) -> list[dict[str, Any]]:
-        """Should perform the actual search request.
-
-        :param args: arguments to be used in the search
-        :param kwargs: keyword arguments to be used in the search
-        :return: list containing the results from the provider in json format
-        """
-        return QueryStringSearch.do_search(self, *args, **kwargs)
