+++ setup.cfg
@@ -156,7 +156,11 @@ console_scripts =
 eodag.plugins.api =
     UsgsApi = eodag.plugins.apis.usgs:UsgsApi [usgs]
     EcmwfApi = eodag.plugins.apis.ecmwf:EcmwfApi [ecmwf]
-    EcmwfGroupApi = eodag.plugins.apis.ecmwf_group:EcmwfGroupApi [ecmwf]
+    EcmwfGroupApi = eodag.plugins.apis.ecmwf_group:EcmwfGroupApi
+    CopEcmwfGroupApi = eodag.plugins.apis.ecmwf_group:CopEcmwfGroupApi
+    WekeoEcmwfGroupApi = eodag.plugins.apis.ecmwf_group:WekeoEcmwfGroupApi
+    DedtEcmwfGroupApi = eodag.plugins.apis.ecmwf_group:DedtEcmwfGroupApi
+
 eodag.plugins.auth =
     GenericAuth = eodag.plugins.authentication.generic:GenericAuth
     HTTPHeaderAuth = eodag.plugins.authentication.header:HTTPHeaderAuth
@@ -189,7 +193,6 @@ eodag.plugins.search =
     StaticStacSearch = eodag.plugins.search.static_stac_search:StaticStacSearch
     ECMWFSearch = eodag.plugins.search.build_search_result:ECMWFSearch
     MeteoblueSearch = eodag.plugins.search.build_search_result:MeteoblueSearch
-    WekeoECMWFSearch = eodag.plugins.search.build_search_result:WekeoECMWFSearch
     DataRequestSearch = eodag.plugins.search.data_request_search:DataRequestSearch
     CreodiasS3Search = eodag.plugins.search.creodias_s3:CreodiasS3Search
     CopMarineSearch = eodag.plugins.search.cop_marine:CopMarineSearch
